Clean up the arguments to sock_bufferize().
Make the first argument a void * - it can point to anything, not just,
for example, a character string.  It's not a "buffer", it's just data,
whatever that might be, so call it "data", not "buffer".

Rename "tempbuf" to "outbuf" to indicate that it's a buffer location
into which data is copied.

In English, stuff would be contained in "data", not contained into
"data" - into suggests that stuff is being copied into it by the
routine, which isn't the case - and stuff isn't copied in "outbuf", it's
copied into "outbuf".

Update the comment for sock_bufferize() to reflect all this.

(cherry picked from commit 316cc596afbabbfd97d10473eb6c1eb72f2c5a73)