rpcapd: clean up the code to fill in the struct rpcap_findalldevs_if.
Make findalldevs_if the on-the-stack struct rpcap_findalldevs_if, and
fill it in directly.  Once it's completely filled in, check that it fits
in the output buffer and copy it to the output buffer using
sock_bufferize().

This makes the code a little easier to figure out.

(cherry picked from commit 95ab2d72bbb979fd65c6536bf6796c17c5f7404d)