configure: AC_TRY_COMPILE() generates a return; don't add one.
If we pass [return 0] to AC_TRY_COMPILE(), the test program it compiles
has two "return 0;" statements in a row, and one of the -W flags we
tests reports a warning for that.

We were testing whether a -W flag is supported by checking the standard
error of the compiler to see if *any* error/warning messages are
generated, and treating the flag as unsupported if any are, that meant
that -Wunreachable-code-return was be treated as unsupported even though
it *is* supported.

This should fix that.  (That flag isn't currently used here, but it is
used for tcpdump, and we want to make this change and the previous
change in tcpdump as well.  I'm so glad autoconf makes this all so
difficult to do correctly....)

(cherry picked from commit 8336c296aff1e7cd0172704ae0e11dd2dc275383)