Fix compiling on Solaris 11.4/AMD64.
In gencode.h and pcap/bpf.h include necessary headers in order not to
depend on other files to do so.

In gencode.c, grammar.c (via grammar.y.in) and scanner.c (via scanner.l)
include gencode.h and grammar.h as early as it takes for the token enum
identifiers not to clash with the named constants defined in system
headers. This fixes the following error with Clang (also with GCC and a
more vague message):

./grammar.h:116:5: error: expected identifier
    ESP = 326,
/usr/include/sys/regset.h:86:14: note: expanded from macro 'ESP'
 #define ESP             7

In build.sh add an exemption rule for the remaining warnings. Now
libpcap can complete the default matrix build (4 rounds with GCC 7.3.0
and 4 rounds with Clang 6.0.0).