Handle XL C a bit more like the other compilers.
The first test is "is this someting that's not XL C", as is the case for
other compilers; that test is done as "are both __xlC__ and __ibmxl__
undefined".  If either of them are defined, use __ibmxl_version__ if
__ibmxl__ is defined, and use __xlC__ otherwise.

That makes it a bit easier to read.

(cherry picked from commit 0ccc80aba83744762181205d5c9b1ff73a7abc2c)