linux: fix check for CAN/CAN FD packets in the direction check routine.
The protocol field is big-endian, not host-endian (it's big-endian in an
sk_buff, and that makes it all the way to userland), so convert it to
host order before comparing with the CAN and CAN FD protocol types.

This addresses part of GitHub issue #1051.

(cherry picked from commit acc5cb90d78e467ccaa24953f6a729c623a99f25)