linux: clean up possibly-uninitialized data in the CAN header.
If the protocol is CAN FD, clear out all bits in the flags field other
than the defined ones, and clear out the two reserved fields.

If the protocol is *not* CAN FD, clear the CAN FD flag, to make sure the
packet is interpreted as classic CAN, not CAN FD.

This is done just in case the kernel or driver put uninitialized junk in
the reserved bits of the CAN header, as has apparently happened in at
least some LINKTYPE_CAN_SOCKETCAN captures, to allow programs reading
those captures to attempt to distinguish "CANFD_FDF is set because the
frame is an FD frame" from "CANFD_FDF is set because the fd_flags field
contains uninitialized junk".

(cherry picked from commit 436179a5f5ff3924ee7285777777a4f71372f654)