Fix pcap_dispatch() to return number of packets processed even at EOF.
If it successfully read one or more packets and then got an EOF, it
would return 0 rather than the number of packets it processed.

While we're at it, change the next_packet_op routines to return 0 on EOF
and 1 on success, rather than vice versa, so that it's a bit clearer.

(cherry picked from commit a37101ac1502b870a57257bbb9900b469dc2caf2)