Improve bpf.h header conflict handling
We carefully check for various operating system include guards to ensure
that we include either the OS header or our own.
This works fine if the OS header is included first, but if we include
our own first and then the OS header we run into duplicate definitions.

This fixes part of the build issues caused by recent FreeBSD including
<net/bpf.h> from <net/if_pflog.h>.