Have a separate DIAG_OFF_ for pointer-to-integer narrowing.
Use that for pcap_fileno() on Windows; that API shouldn't exist on
Windows, and should never be used, but WinPcap offered it, so we're
stuck with it.  At least try to prevent compilation errors when libpcap
is built.

This should fix the warnings for pcap_fileno() in GitHub issue #1026.

(backported from commit ea1928b35bddea69b8843161f2f8dbff54c0d74e)