pcap_handle: fix deprecation warning.
Don't tell people to use a routine that doesn't exist; instead, if they
need it, tell them to request it ("request" includes "give us a reason
to believe that they really need it" - what operations do you expect to
do on the HANDLE in question?  Note that pcap_getevent() is what you
want if you want to wait for an event on it.).