Revert "findalldevstest: temporary test to see if this squelches Valgrind warnings."
This reverts commit 778f2c0e806090dc524627161bca826cc1d061a2.

Yes, Illumos (and probably Solaris) may fflush() all the standard I/O
streams before exiting, but it doesn't fclose() them, so stuff might not
be freed before exiting (but it should be accessible from stdin, stdout,
or stderr, so it's at worst a "still reachable" issue - what gives?).

There *appears* to be a suppressor for that in the Solaris valgrind, but
it doesn't seem to be working.