pcap-usb-linux: clean up the calculation of the packets-to-fetch value.
This should produce the same results as the previous code, but it's a
little clearer what's going on.  Add comments to explain what's being
done (from which one can infer that it produces the same results).

Add more comments explaining other things in the code and noting a
potential issue with allowing pcap_breakloop() to break out of the
packet-processing inner loop.