linux: fix a routine name.
iface_get_ts_info() does use an ethtool ioctl whose name ends with
GET_TS_INFO to fetch information about the adapter's time stamping
support, but the goal is to get the types of time stamping we support.
Rename it to iface_get_ts_types() to make it clearer what it does.

Update a comment to express a hope that said ioctl has been fixed not to
lie as much as it did when I sent that message to linux-netdev back in
2016 (the ixgbe driver appears to have been fixed).