NPF: move the "fetch time stamp types" code to its own routine.
That cleans the code up a bit.

We also have a version of that routine that does nothing, and use it if
we *don't* have PacketGetTimestampModes().  That removes an #ifdef from
pcap_create_interfaces().

We can just use pcap_close() if get_ts_types() fails, as its
cleanup_op() was set to the default value of pcap_cleanup_live_common(),
which will free up the time stamp type list if it's been allocated.

This all matches how we do this in pcap-linux.c.