configure: use PKG_CHECK_MODULES to run pkg-config.
Change PKG_CHECK_MODULES not to fail if pkg-config isn't found, because
it's not present by default on all of the platforms we support, and we
don't want to require it to be installed.

Clean up some of its text to use the name of the package, rather than
the prefix used for its variables.

Use it in place of the code we had before to run pkg-config.