configure: get rid of the attempt to auto-run PKG_PROG_PKG_CONFIG.
As the comment said:

    Please remember that m4 expands AC_REQUIRE([PKG_PROG_PKG_CONFIG])
    only at the first occurrence in configure.ac, so if the first place
    it's called might be skipped (such as if it is within an "if", you
    have to call PKG_CHECK_EXISTS manually

This is fragile as heck and, in fact, the first place we call any macro
that uses AC_REQUIRE([PKG_PROG_PKG_CONFIG]) *is* inside an "if".

Explicitly call PKG_PROG_PKG_CONFIG up front in configure.ac, and get
rid of the attempts to automatically run it; those attempts buy us
nothing.