configure: clean up searching for OpenSSL/libressl.
First, try using pkg-config, and only then fall back on other
mechanisms.

When using pkg-config, if we have Homebrew installed, set
PKG_CONFIG_PATH to try to find the openssl.pc file installed by Homebrew
before running pkg-config.  Don't wire in the path for Homebrew, as it's
different on macOS-for-x86 and macOS-for-ARM; instead, fetch it using
"brew --prefix".

That should obviate the need to manually look for the Homebrew version,
so don't do that.

It should also obviate the need to do the weird check for "we have the
macOS-supplied library, but Xcode doesn't supply headers; we also have a
third-party version installed, but we end up building with the
third-party version's headers and the macOS-supplied library, which
doesn't work"; note that.

(cherry picked from commit ccd2826a9c6c7c3a53f88c2d9901d17e4d5fe740)