configure: clean up checking for openssl with Homebrew and with nothing.
For Homebrew, use the brew command to find out where to search with
pkg-config and, if we don't have pkg-config or don't find it with
pkg-config, use it to find out if OpenSSL/libressl is installed and, if
so, use it to fid out what the appropriate include and lib directories
are.

If that fails, try looking in /usr/local if both /usr/local/include and
/usr/local/lib exist, by adding -I and -L flags for those directories.
That makes sure we don't get caught by the "we have a third-party
version installed under /usr/local and also have the no-headers version
from macOS" problem; the explicit -I and -L flags force the compiler and
linker to look there forst, so we get the thrid-party library rather
than the macOS-supplied library (which doesn't work with the third-party
headers).

If *that* fails, just try without any -I or -L flags, to see if the
system version is usable.

(cherry picked from commit 5dc08595548b5a0323c2f75b828a16d544fdbd8e)