npf: don't have pcap_create() fail if the device doesn't exist.
Also, don't have it fail if the user doesn't have permission to open it.

To quote the comment (which opens with a paragraph taken directly from
the equivalent code in pcap-linux.c):

If we can't open the device now, we won't be able to later, either.

If the error is something that indicates that the device doesn't exist,
or that they don't have permission to open the device - or perhaps that
they don't have permission to get a list of devices, if
PacketOpenAdapter() does that - the user will find that out when they
try to activate the device; just return an empty list of time stamp
types.

Treating either of those as errors will, for example, cause "tcpdump -i
<number>" to fail, because it first tries to pass the interface name to
pcap_create() and pcap_activate(), in order to handle OSes where
interfaces can have names that are just numbers (stand up and say hello,
Linux!), and, if pcap_activate() fails with a "no such device" error,
checks whether the interface name is a valid number and, if so, tries to
use it as an index in the list of interfaces.

That means pcap_create() must succeed even for interfaces that don't
exist, with the failure occurring at pcap_activate() time.