pcap-usb-linux: fix scan for usbmon devices.
The directory to scan is just /dev, which is wired into the udev source,
at least in the systemd source tree.  Just scan that directory, rather
than using using udevinfo to find out the directory.  (If there is ever
a reason to look anywhere there than /dev for the usbmon devices - and,
if so, there had better be a good reason, as stuffing special files
anywhere other than /dev on any UN*X, including Linux, will break
programs that use /dev/null or /dev/tty, among other well-known devices
- then provide a mechanism that works at *run time*, so that it works
even for cross-builds libpcap, and that works regardless of whether the
machine on which the built version of libpcap has udev and whether it
has udevinfo or udevadm or whatever.)