Make the error message for PCAP_ERROR_PERM_DENIED a little more open-ended.
PCAP_ERROR_PERM_DENIED can be returned for reasons other than "you don't
have any permission to capture on that device", e.g. it might be "you
have permission to capture on it, but not to capture on it with hardware
time stamps (as on Linux) or to capture in monitor mode or to...".

This leaves it more open-ended, in case we add support for
device-dependent or platform-dependent configuration parameters and some
of them require more privileges than just capturing with the default
setting does, without having to keep adding more PERM_DENIED error
codes.

This, along with some other changes, may help debug problems such as the
one in issue #1098.

(cherry picked from commit dcb90c46c20cb3483c4e53484adb3264dc7902f0)