configure, cmake: only check for DPDK using pkg-config.
We only try to find DPDK using pkg-config; DPDK is *SO* complicated -
DPDK 19.02, for example, has about 117(!) libraries, and the precise set
of libraries required has changed ver time - so attempting to guess
which libraries you need, and hardcoding that in an attempt to find the
libraries without DPDK, rather than relying on DPDK to tell you, with a
.pc file, what libraries are needed, is *EXTREELY* fragile and has
caused some bug reports, so we're just not going to do it.

If that causes a problem, the only thing we will do is accept an
alternative way of finding the appropriate library set for the installed
version of DPDK that is as robust as pkg-config (i.e., it had better
work as well as pkg-config with *ALL* versions of DPDK that provide a
libdpdk.pc file).

Also, try to give as much information to the user as possible if we fail
to find a usable DPDK.