INSTALL.md: update and expand. [skip ci]
Mention CMake builds as well as autotools bulds.

Use a bulleted list for build steps with autotools and CMake.

Note that both the configure script and CMake support options to
configure the build.

Note that we support an "uninstall" target as well as an "install"
target.

Strongly recommend *against* doing the build as root.

Update the note about the "cannot determine packet capture interface"
message to give the current text and to note that CMake also issues it,
and to indicate the options to use with autotools and CMake to set the
interface from the command line.

Note that CMake doesn't check whether Lex/Flex or YACC is adequate to
build libpcap.