Finddpdk: don't do stuff requiring 3.19 or later on earlier CMakes.
There's a bunch of "imported targets" stuff in the Finddpdk CMake
module, but support for that wasn't added until 3.19.  Don't do it if
we're using a version of CMake prior to 3.19.

(No, we're not going to just require 3.19 or later.  See the comment I
added.)

(cherry picked from commit ee9097ae21315c3f282d86006f439770c832eb80)