linux: don't bother testing for PACKET_AUXDATA support.
We require Linux 2.6.27 or later, and that version has PACKET_AUXDATA.

Also update a question to reflect that 1) the build will fail before
that point if it's not building for 2.6.27 or later, so auxiliary data
is guaranteed to be supported and 2) it appears that, for memory-mapped
capture - which is all we support - the auxiliary data is provided
regardless of whether it's enabled.

This was noticed in issue #1105; the PACKET_AUXDATA support wasn't
getting compiled in.

(cherry picked from commit 3a7bb4ab43d14ba61eaf07921dd2e8f875281657)