linux: rename a routine to make its function clearer.
Its function is to set up the socket for capture, so call it
setup_socket().  "pf_packet" in "activate_pf_packet()" iss redundant, as
the only type of socket we now support is PF_PACKET (we dropped support
for systems without PF_PACKET sockets a while ago), and it doesn't do a
full activation, it just does some socket setup.