Close the eventfd if we are non-blocking
The eventfd is used to break out of a poll() before it
times out, used by pcap_breakloop().  If we are non-blocking,
then the eventfd is never needed, so we close it.  (And
open a new eventfd if we switch to blocking).

(cherry picked from commit 024fd9717d43835f3e08b3146714532768910d3c)