Improve some protocol details in pcap-filter(7).
Address point 2 and progress point 5 of GH bug report #326.

In the "ip proto protocol" section omit the "icmp6" protocol name,
explain the implicit involvement of getprotobyname() and give two
OS-specific names for IGRP.  In the "ip6 proto protocol" section give
"ipv6-icmp" as the proper protocol name (not a keyword).  In the index
operation description add "icmp6" to the list of supported protocol
layers.

Add two entries with the meaning of "icmp" and "icmp6" abbreviations.
In the list of abbreviations for "proto \protocol" remove "icmp" because
it never was and should not be the implemented behaviour; add "pim",
"ah" and "esp" because that has been the implemented behaviour since
commit 7fe3c11 in 1999.

Reformat and sort some lists alphabetically, update the timestamp.

(cherry picked from commit db3b00e9a0875edea808734640d84e035ae11357)