Remove the unused CHASE_CHAIN props.
CHASE_CHAIN is not correlated with the build process, it has been
undefined since it was introduced in commit 7fe3c11 in 1999, possibly as
a leftover from earlier prototypes.  If defined, the code still compiles
fine, but quietly changes the behaviour in meaningful ways, which is a
wrong thing to do.  Remove the macro and the associated dead code.

A practical side effect of this change is that gen_protochain() now can
be conditional solely on NO_PROTOCHAIN, which simplifies getting the
latter right to solve the build problem stated in GH #852.