Linux USB: make the capture file length fixup more conservative.
When retroactively fixing up capture file on-the-network lengths for
memory-apped Linux USB captures, only fix up the on-the-network length
if it has the value that would be expected from the old capture code.

In the fixup code, make sure it's both >= the proper length, as best we
can calculate it, and the captured length.