Refine "relop" notes in pcap-filter(7). [skip ci]
Say "expr1 relop expr2" to emphasize the operands are most likely
different expressions.  Split a long sentence for clarity.  Consistent
with the rest of the document, show operators in bold.  Use the standard
set notation.  Move the length operator to a paragraph of its own so the
section is easier to follow.  In the mod/xor support note provide more
OS particulars.