Improve handling of C++.
autotools:

Determine the size of a void *; we don't use it now except for the
following test, but we will be using it in the future to determine
whether we're doing a 32-bit or 64-bit build, so that we can bludgeon
pkg-config into finding the right versions of libraries.

Check to make sure that the C and C++ compiler don't generate code for
different data models; if they do, it means we can't build C and C++
code and combine them, and also means that CMAKE_SIZEOF_VOID_P won't
"reflect reality" because there's no single reality to reflect, and we
won't be able to use it to determine whether we're doing a 32-bit or
64-bit build (which we may have to do in the future to make pkg-config
work correctly on some platforms).  There's unlikely to be a mismatch in
Haiku, but we do this anyway for future-proofing.

CMake:

Only check for a C++ compiler on Haiku.

Check to make sure that the C and C++ compilers don't generate code for
different data models.

This is a first step for issue #1112.

(cherry picked from commit 914e40f62f3382256cc27d5e171661f92f840f81)