On Solaris, tweak PKG_CONFIG_PATH as necessary for the build bitwidth.
For 32-bit builds, make sure that, if /usr/lib/amd64/pkgconfig is in the
path, we put /usr/lib/pkgconfig before it, so that we find the .pc files
for 32-bit versions of packages.

For 64-bit builds, make sure that, if /usr/lib/pkgconfig is in the path,
we put /usr/lib/amd64/pkgconfig before it and, if it's not in the path,
we have /usr/lib/amd64/pkgconfig at the end of the path, so that we find
the .pc files for 64-bit versions of packages.

The dbus package makes this necessary.

This should fix issue #1112.