bpf: make the BPF device number unsigned.
It's unlikely that we'll try 2^31-1 devices without success and overflow
to 2^31, but this is cleaner (and BPF devices won't have negative device
numbers).

That means the buffer for the device name doesn't need to leave room for
the minus sign; that's what the compiler was warning about.

(cherry picked from commit 7610a61d72c818067a1e51f150b81ee7d489a577)