CHANGES: move some changes that affect building. [skip ci]
"Source code" is mostly for changes that won't be seen by users who just
download, configure, and build because they want a library binary; the
changes in question affect stuff those users will see, so move them to
"Building and testing".

Add a GitHub issue number to one of them.