appveyor: remove MinGW-w64, add VS 2022.
Npcap's latest SDK adds SAL annotations to Packet32.h, and that doesn't
seem to work with the version of MinGW-w64 that we have.  That's the
latest version that AppVeyor claim to support, so just drop it for now.
If a newer version supports the annotations, and either AppVeyor adds
support for that newer version or if we can just switch to Cirrus
CI and use that newer version, we can bring it back.

Add VS 2022.  This may require more tweakage.

Try to prevent at least some of the non-Windows builders from wasting
time with this:

[skip ci]