rpcap: improve error messages for host and port resolution errors.
If we don't want a particular port nuber in a sock_initaddress() call,
pass NULL rather than "0".  If the service name parameter passsed to
sock_initaddress() is NULL, pass "0" as the service name parameter to
getaddrinfo().

Have get_gai_errstring() precede the host/port name information with an
indication as to whethe it's a host name, port name, or host name and
port name.  Don't say "host name" for EAI_NONAME; rely on the
description get_gai_errstring() provides.  If there's only a port
number, don't preceded it with ":" in get_gai_errstring().

This makes the error message reported if a host and port are provided
not say that the host name couldn't be resolved, because it could be a
problem with the port name (sadly, getaddinfo() doesn't indicate which
is the one with the problem).

It also makes the error message reported if only a port is provided not
say that it's a problem with the host name or show the "host name" as
":<port>".

(cherry picked from commit 33cf6fb70a13a982d70f6a5e5e63aa765073c8e8)