rpcapd: add a -t flag to specify the data channel port.
If t is specified, instead of allowing the OS to choose a port number
for the data channel, rpcapd forces use of the specified port number.

This allows a port for which firewalls between the client and server
will allow traffic to pass.

This was taken from GitHub pull request rpcapd-linux/rpcapd-linux#4
(https://github.com/rpcapd-linux/rpcapd-linux/pull/4).

This should address #1120.