pcap-config: don't assume the system library directory is /usr/lib.
Instad of checking whether $libdir is /usr/lib in order to decide
whether to bother adding a -L$libdir flag or not, check whether $prefix
is /usr.

We must check $prefix, as $libdir isn't necessarily /usr/lib in this
case - for example, Linux distributions for 64-bit platforms that also
provide support for binaries for a 32-bit version of the platform may
put the 64-bit libraries, the 32-bit libraries, or both in directories
other than /usr/lib.