cmake: fix up generation of libpcap.pc and pcap-config.
For all dependencies found by pkg-config, directly ask pkg-config for
--libs and --libs-static; don't rely on anything from CMake, as CMake
reduces the information content of what *it* gets from --libs and
--libs-static to a level insufficient to reconstruct what pkg-config
gave it.

For dependencies not found by pkg-config, slice the library name up and
use that to construct -L and -l flags.

Make sure we get enough information to fill in LIBS, as that's what now
gets reported by pcap-config --libs --static-pcap-only.