cmake: just show whether the build is 32-bit or 64-bit.
32-bit vs. 64-bit builds seems to be the issue with a significant
number of annoying problems, and we can extract *that* information from
CMake fairly reliably, unlike, say, the target architecture.  (building
with a different bit width from the platform's bit width is not
cross-compiling, so CMake just set CMAKE_SYSTEM_PROCESSOR to the same
value as CMAKE_HOST_SYSTEM_PROCESSOR, which, on UN*Xes, is the output of
"uname -p".  On 64-bit Solaris, that may be a name reflecting the
*32-bit version* of the instruction set, which is incorrect if we're
doing 64-bit builds, e.g. with GCC or with "cc -m64" - no, the default
for Sun^WOracle Studio's compiler is *32-bit*, even on 64-bit platforms.
Go figure.)