cmake: fix check for libibverbs with Sun's C compiler.
For unknown reasons, check_include_file() doesn't just attempt to
compile a test program that includes the header in question, it also
attempts to link it.

For unknown reasons, at least some of the static inline functions
defined in infiniband/verbs.h are not inlined by the Sun^WOracle Studio
C compiler, so the compiler generates code for them as part of the
object code resulting from compiling the test program.  At lest some of
those functions call routines in -libverbs, so, in order to keep the
compile and link from failing, even though the header file exists and is
usable, we need to link with -libverbs.