DllMain() isn't required in a DLL, so remove it.
It used to set up a critical section intended to protect pcap_compile()
against simultaneous use in more than one thread, but that is no longer
necessary, given that pcap_compile() is now thread-safe; the code to set
up the critical section was removed when the first changes to make the
compiler thread-safe were made.

This should fix issue #1127.