pcap-filter(7): Refine markup of a few lists. [skip ci]
Put each item on a separate line and prevent hyphenation, this way the
source is more convenient to edit and the output isn't as much distorted
when rendered using man.

Do not reorder any items on this occasion: at least in the ICMP and
ICMPv6 lists the ordering exactly matches the source code and makes it
easy to verify the contents.

The ICMPv6 types list comprises different values of the same protocol
field rather than different protocol fields, fixup the wording to
reflect that consistently with the ICMP types list just above.