Remove doc/README.capture-module
The online link in the chapter 8.1 of Wireshark Developer's Guide was
on the master branch. It is therefore not necessary to keep this file
on a release branch.

The link was recently updated. But we will keep the file for several
months in the master branch for readers of offline formats.